Make SCollection WithResource DoFn available without import tax
