+++ scio-core/src/main/scala/com/spotify/scio/transforms/package.scala
@@ -88,106 +88,6 @@ package object transforms {
         .flatMap(identity)
   }
 
-  class CollectFnWithResource[T, U, R] private[transforms] (
-    resource: => R,
-    resourceType: ResourceType,
-    pfn: PartialFunction[(R, T), U]
-  ) extends DoFnWithResource[T, U, R] {
-    def getResourceType: ResourceType = resourceType
-    def createResource: R = resource
-
-    val isDefined: ((R, T)) => Boolean = ClosureCleaner.clean(pfn.isDefinedAt) // defeat closure
-    val g: PartialFunction[(R, T), U] = ClosureCleaner.clean(pfn)
-    @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit =
-      if (isDefined((getResource, c.element()))) {
-        c.output(g((getResource, c.element())))
-      }
-  }
-
-  class MapFnWithResource[T, U, R] private[transforms] (
-    resource: => R,
-    resourceType: ResourceType,
-    f: (R, T) => U
-  ) extends DoFnWithResource[T, U, R] {
-    def getResourceType: ResourceType = resourceType
-    def createResource: R = resource
-
-    val g: (R, T) => U = ClosureCleaner.clean(f)
-    @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit =
-      c.output(g(getResource, c.element()))
-  }
-
-  class FlatMapFnWithResource[T, U, R] private[transforms] (
-    resource: => R,
-    resourceType: ResourceType,
-    f: (R, T) => TraversableOnce[U]
-  ) extends DoFnWithResource[T, U, R] {
-    def getResourceType: ResourceType = resourceType
-    def createResource: R = resource
-
-    val g: (R, T) => TraversableOnce[U] = ClosureCleaner.clean(f)
-    @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit = {
-      val i = g(getResource, c.element()).toIterator
-      while (i.hasNext) c.output(i.next())
-    }
-  }
-
-  class FilterFnWithResource[T, R] private[transforms] (
-    resource: => R,
-    resourceType: ResourceType,
-    f: (R, T) => Boolean
-  ) extends DoFnWithResource[T, T, R] {
-    def getResourceType: ResourceType = resourceType
-    def createResource: R = resource
-
-    val g: (R, T) => Boolean = ClosureCleaner.clean(f)
-    @ProcessElement def processElement(c: DoFn[T, T]#ProcessContext): Unit =
-      if (g(getResource, c.element())) {
-        c.output(c.element())
-      }
-  }
-
-  implicit class SCollectionWithResourceFunctions[T](private val self: SCollection[T])
-      extends AnyVal {
-
-    /**
-     * Return a new [[SCollection]] by applying a function that also takes in a resource and
-     * `ResourceType` to all elements of this SCollection.
-     */
-    def mapWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-      fn: (R, T) => U
-    ): SCollection[U] =
-      self.parDo(new MapFnWithResource(resource, resourceType, fn))
-
-    /**
-     * Filter the elements for which the given `PartialFunction` that also takes in a resource and
-     * `ResourceType` is defined, and then map.
-     */
-    def collectWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-      pfn: PartialFunction[(R, T), U]
-    ): SCollection[U] =
-      self.parDo(new CollectFnWithResource(resource, resourceType, pfn))
-
-    /**
-     * Return a new [[SCollection]] by first applying a function that also takes in a resource and
-     * `ResourceType` to all elements of this SCollection, and then flattening the results.
-     */
-    def flatMapWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-      fn: (R, T) => TraversableOnce[U]
-    ): SCollection[U] =
-      self.parDo(new FlatMapFnWithResource(resource, resourceType, fn))
-
-    /**
-     * Return a new [[SCollection]] containing only the elements that satisfy a predicate that
-     * takes in a resource and `ResourceType`
-     */
-    def filterWithResource[R](resource: => R, resourceType: ResourceType)(
-      fn: (R, T) => Boolean
-    ): SCollection[T] =
-      self.parDo(new FilterFnWithResource(resource, resourceType, fn))(self.coder)
-
-  }
-
   /**
    * Enhanced version of [[com.spotify.scio.values.SCollection SCollection]] with custom
    * parallelism, where `parallelism` is the number of concurrent `DoFn` threads per worker
